Move some code into utils and fix the executable
